Integrated Stripe payment gateway

+++ backend/shoppingApp/views.py
@@ -1,8 +1,8 @@
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.views import APIView
-from .models import Category, Product, Order, OrderItem, Payment
-from .serializer import CategorySerializer, ProductSerializer, OrderSerializer, OrderItemSerializer, PaymentSerializer
+from .models import *
+from .serializer import *
 from core.models import UserCustomModel as User
 from core.serializer import UserSerializer 
 import json
@@ -145,33 +145,33 @@ class OrderListCreateAPIView(APIView):
             return Response({"success": False, "error": str(e), "status": 400}, status=status.HTTP_400_BAD_REQUEST)
 
 class OrderDetailAPIView(APIView):
-    permission_classes=[IsAuthenticated]
+    # permission_classes=[IsAuthenticated]
     def get(self, request, pk):
-        try:
-            order = Order.objects.get(user=request.user)
-        except Order.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
-        serializer = OrderSerializer(order)
-        return Response(serializer.data)
-
-    def patch(self, request, pk):
         try:
             order = Order.objects.get(pk=pk)
         except Order.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
+            return Response({"success":False,"error":"Not Found","status":404},status=status.HTTP_404_NOT_FOUND)
+        serializer = OrderSerializer(order)
+        return Response({"success": True, "data": serializer.data, "status": 200})
+
+    def patch(self, request, pk):
+        order = Order.objects.filter(pk=pk).first()
+        if not order:  # Check if the order exists
+            return Response({"success":False,"error":"Not Found","status":404},status=status.HTTP_404_NOT_FOUND)
         serializer = OrderSerializer(order, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return Response({"success": True, "data": serializer.data, "status": 200})
+        return Response({"success": False, "error": serializer.errors, "status": 400}, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk):
-        try:
-            order = Order.objects.get(pk=pk)
-        except Order.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
+      
+        
+        order = Order.objects.filter(pk=pk).first()
+        if not order:  # Check if the order exists
+            return Response({"success":False,"error":"Not Found","status":404},status=status.HTTP_404_NOT_FOUND)
         order.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response({"success":True,"data":"Order Deleted Successfully","status":204},status=status.HTTP_204_NO_CONTENT)
 
 # OrderItem API Views
 class OrderItemListCreateAPIView(APIView):
@@ -234,44 +234,44 @@ class OrderItemDetailAPIView(APIView):
         order_item.delete()
         return Response({"success":True,"data":"Order Item Deleted Successfully","status":204},status=status.HTTP_204_NO_CONTENT)
 
-# Payment API Views
-class PaymentListCreateAPIView(APIView):
-    def get(self, request):
-        payments = Payment.objects.all()
-        serializer = PaymentSerializer(payments, many=True)
-        return Response(serializer.data)
+# # Payment API Views
+# class PaymentListCreateAPIView(APIView):
+#     def get(self, request):
+#         payments = Payment.objects.all()
+#         serializer = PaymentSerializer(payments, many=True)
+#         return Response(serializer.data)
 
-    def post(self, request):
-        serializer = PaymentSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+#     def post(self, request):
+#         serializer = PaymentSerializer(data=request.data)
+#         if serializer.is_valid():
+#             serializer.save()
+#             return Response(serializer.data, status=status.HTTP_201_CREATED)
+#         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-class PaymentDetailAPIView(APIView):
-    def get(self, request, pk):
-        try:
-            payment = Payment.objects.get(pk=pk)
-        except Payment.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
-        serializer = PaymentSerializer(payment)
-        return Response(serializer.data)
+# class PaymentDetailAPIView(APIView):
+#     def get(self, request, pk):
+#         try:
+#             payment = Payment.objects.get(pk=pk)
+#         except Payment.DoesNotExist:
+#             return Response(status=status.HTTP_404_NOT_FOUND)
+#         serializer = PaymentSerializer(payment)
+#         return Response(serializer.data)
 
-    def patch(self, request, pk):
-        try:
-            payment = Payment.objects.get(pk=pk)
-        except Payment.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
-        serializer = PaymentSerializer(payment, data=request.data, partial=True)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+#     def patch(self, request, pk):
+#         try:
+#             payment = Payment.objects.get(pk=pk)
+#         except Payment.DoesNotExist:
+#             return Response(status=status.HTTP_404_NOT_FOUND)
+#         serializer = PaymentSerializer(payment, data=request.data, partial=True)
+#         if serializer.is_valid():
+#             serializer.save()
+#             return Response(serializer.data)
+#         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def delete(self, request, pk):
-        try:
-            payment = Payment.objects.get(pk=pk)
-        except Payment.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
-        payment.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+#     def delete(self, request, pk):
+#         try:
+#             payment = Payment.objects.get(pk=pk)
+#         except Payment.DoesNotExist:
+#             return Response(status=status.HTTP_404_NOT_FOUND)
+#         payment.delete()
+#         return Response(status=status.HTTP_204_NO_CONTENT)
